requirements.txt
tweepy
requests
pillow
pip
aggdraw
fpdf2
uuid